fix most lint errors
