+++ addons/viewport/manager.js
@@ -1,2 +1,3 @@
 const manager = require('./dist/manager');
+
 manager.init();
