+++ addons/viewport/src/manager.jsx
@@ -5,19 +5,18 @@ import { Panel } from './components/Panel';
 
 const ADDON_ID = 'storybook-addon-viewport';
 const PANEL_ID = `${ADDON_ID}/addon-panel`;
-const EVENT_ID = `${ADDON_ID}/addon-event`;
 
 function init() {
-    addons.register(ADDON_ID, api => {
-        const channel = addons.getChannel();
+  addons.register(ADDON_ID, api => {
+    const channel = addons.getChannel();
 
-        addons.addPanel(PANEL_ID, {
-            title: 'Viewport',
-            render() {
-                return <Panel channel={channel} api={api} />;
-            }
-        });
+    addons.addPanel(PANEL_ID, {
+      title: 'Viewport',
+      render() {
+        return <Panel channel={channel} api={api} />;
+      },
     });
+  });
 }
 
-export { init }
+export { init };
